Main Class
This program asks for number of test cases and time of meeting, number of friends and then the time range for which friends are available and then prints 1 if friend is available for meeting and 0 if not. It can also perform sub tasks of validating 12 hour format time and change 12 hour to 24 hour time fromat.